ТЗ-10 fix #1
