+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -13,7 +13,4 @@ public interface FilmStorage {
     Film addFilm(Film film);
 
     Film updateFilm(Film film);
-
-    void validateFilm(Film film);
-
 }
