+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -13,7 +13,4 @@ public interface UserStorage {
     User createUser(User user);
 
     User updateUser(User user);
-
-    void validateUser(User user);
-
 }
