+++ src/test/java/ru/yandex/practicum/filmorate/service/FilmServiceTest.java
@@ -17,6 +17,7 @@ import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 import java.time.LocalDate;
 import java.time.Month;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @SpringBootTest
@@ -24,11 +25,10 @@ public class FilmServiceTest {
 
     FilmStorage filmStorage = new InMemoryFilmStorage();
     UserStorage userStorage = new InMemoryUserStorage();
-    FilmService filmService = new FilmService(filmStorage, userStorage);
-    FilmController filmController = new FilmController(filmStorage, filmService);
-
     UserService userService = new UserService(userStorage);
-    UserController userController = new UserController(userStorage, userService);
+    FilmService filmService = new FilmService(filmStorage, userService);
+    FilmController filmController = new FilmController(filmService);
+    UserController userController = new UserController(userService);
 
     @Test
     void getFilmById_returnFilm() {
@@ -51,6 +51,102 @@ public class FilmServiceTest {
     }
 
     @Test
+    void getAllFilms_returnAllFilmsListSize() {
+        filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
+        filmController.addFilm(new Film("Терминатор 2", "Шварцнеггер хороший", LocalDate.of(1990, Month.JANUARY, 1), 100));
+
+        Assertions.assertEquals(2, filmController.getAllFilms().size());
+    }
+
+    @Test
+    void addFilm_nullFilm() {
+        ValidationException nonuser = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(null)
+        );
+        Assertions.assertEquals("Данные фильма не переданы", nonuser.getMessage());
+    }
+
+    @Test
+    void addFilm_returnAllId() {
+        Film film1 = filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
+        Film film2 = filmController.addFilm(new Film("Терминатор 2", "Шварцнеггер хороший", LocalDate.of(1990, Month.JANUARY, 1), 100));
+        Film film3 = filmController.addFilm(new Film("Терминатор 3", "Шварцнеггер никакой", LocalDate.of(2003, Month.JANUARY, 1), 100));
+
+        Assertions.assertEquals(1, film1.getId());
+        Assertions.assertEquals(2, film2.getId());
+        Assertions.assertEquals(3, film3.getId());
+        Assertions.assertEquals("Терминатор 1", film1.getName());
+        Assertions.assertEquals("Шварцнеггер плохой", film1.getDescription());
+        Assertions.assertNotNull(film1.getReleaseDate());
+        Assertions.assertNotNull(film1.getDuration());
+    }
+
+    @Test
+    void addFilm_wrongName() {
+        ValidationException film1 = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(new Film("", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100))
+        );
+        Assertions.assertEquals("Не указано название фильма", film1.getMessage());
+
+        ValidationException film2 = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(new Film(null, "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100))
+        );
+        Assertions.assertEquals("Не указано название фильма", film2.getMessage());
+    }
+
+    @Test
+    void addFilm_tooLongDescription() {
+        String description = "a".repeat(201);
+
+        ValidationException film1 = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(new Film("Терминатор 1", description, LocalDate.of(1984, Month.JANUARY, 1), 100))
+        );
+        Assertions.assertEquals("Размер описания не должен превышать 200 символов", film1.getMessage());
+    }
+
+    @Test
+    void addFilm_wrongRelease() {
+        ValidationException film1 = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1895, Month.DECEMBER, 27), 100))
+        );
+        Assertions.assertEquals("Дата релиза указана некорректна", film1.getMessage());
+    }
+
+    @Test
+    void addFilm_wrongDuration() {
+        ValidationException film1 = assertThrows(
+                ValidationException.class,
+                () -> filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), -10))
+        );
+        Assertions.assertEquals("Продолжительность фильма указана некорректна", film1.getMessage());
+    }
+
+    @Test
+    void updateFilm_returnUpdatedFilm() {
+        filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
+        Film updFilm = new Film("Терминатор 1", "Шварцнеггер чоткий", LocalDate.of(1984, Month.JANUARY, 1), 100);
+        updFilm.setId(1);
+        filmController.updateFilm(updFilm);
+
+        assertEquals(filmController.getFilmById(1), updFilm);
+    }
+
+    @Test
+    void updateFilm_emptyFilmMap() {
+        NotFoundException nonuser0 = assertThrows(
+                NotFoundException.class,
+                () -> filmController.updateFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100))
+        );
+        Assertions.assertEquals("Такой фильм отсутствует", nonuser0.getMessage());
+    }
+
+
+    @Test
     void likeFilm_returnAmountOfLike() {
         userController.createUser(new User("user1@ya.ru", "user1", LocalDate.of(2000, Month.JANUARY, 1)));
         userController.createUser(new User("user2@ya.ru", "user2", LocalDate.of(2000, Month.JANUARY, 2)));
@@ -69,18 +165,17 @@ public class FilmServiceTest {
                 NotFoundException.class,
                 () -> filmController.likeFilm(1, 1)
         );
-        Assertions.assertEquals("Фильм id=1 не найден", nonFilm.getMessage());
+        Assertions.assertEquals("Такой фильм отсутствует", nonFilm.getMessage());
     }
 
     @Test
     void likeFilm_wrongUserId() {
         filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
-
         NotFoundException nonUser = assertThrows(
                 NotFoundException.class,
                 () -> filmController.likeFilm(1, 1)
         );
-        Assertions.assertEquals("Пользователь id=1 не найден", nonUser.getMessage());
+        Assertions.assertEquals("Такой пользователь отсутствует", nonUser.getMessage());
     }
 
     @Test
@@ -101,12 +196,7 @@ public class FilmServiceTest {
     void revokeLikeToFilm_nonExistedLike() {
         userController.createUser(new User("user1@ya.ru", "user1", LocalDate.of(2000, Month.JANUARY, 1)));
         filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
-
-        ValidationException nonLike = assertThrows(
-                ValidationException.class,
-                () -> filmController.revokeLikeToFilm(1, 1)
-        );
-        Assertions.assertEquals("Фильм id=1 не получал лайк от пользователя id=1", nonLike.getMessage());
+        filmController.revokeLikeToFilm(1, 1);
     }
 
     @Test
@@ -134,11 +224,7 @@ public class FilmServiceTest {
 
     @Test
     void getTopPopFilms_emptyList_wrongCount() {
-        ValidationException emptyList0 = assertThrows(
-                ValidationException.class,
-                () -> filmController.getTopPopFilms(1)
-        );
-        Assertions.assertEquals("Список фильмов пуст", emptyList0.getMessage());
+        Assertions.assertEquals(0, filmController.getTopPopFilms(5).size());
 
         filmController.addFilm(new Film("Терминатор 1", "Шварцнеггер плохой", LocalDate.of(1984, Month.JANUARY, 1), 100));
